refactor(tests): Move fixture to `conftest.py`; create util module
Move assertion function for type checking to util module, move
lists and dictionaries of numpy dtypes to util module.
